refactor: add interface for TasksWorkManager
