+++ app/src/main/java/com/example/plantsapp/data/usecase/SignOutUseCaseImpl.kt
@@ -3,7 +3,7 @@ package com.example.plantsapp.data.usecase
 import com.example.plantsapp.di.module.FirebaseQualifier
 import com.example.plantsapp.domain.repository.UserRepository
 import com.example.plantsapp.domain.usecase.SignOutUseCase
-import com.example.plantsapp.presentation.ui.worker.TasksWorkManager
+import com.example.plantsapp.domain.workmanager.TasksWorkManager
 import com.google.android.gms.auth.api.signin.GoogleSignInClient
 import com.google.firebase.auth.FirebaseAuth
 import kotlinx.coroutines.tasks.await
