+++ app/src/main/java/com/example/plantsapp/domain/workmanager/TasksWorkManager.kt
@@ -1,0 +1,8 @@
+package com.example.plantsapp.domain.workmanager
+
+import java.util.Calendar
+
+interface TasksWorkManager {
+    fun startWork(startDate: Calendar)
+    fun cancelAllWork()
+}
