+++ app/src/main/java/com/example/plantsapp/presentation/ui/profile/ProfileViewModel.kt
@@ -9,7 +9,6 @@ import com.example.plantsapp.domain.model.User
 import com.example.plantsapp.domain.repository.UserRepository
 import com.example.plantsapp.domain.usecase.SignOutUseCase
 import com.example.plantsapp.presentation.core.Event
-import com.example.plantsapp.presentation.ui.worker.TasksWorkManager
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import javax.inject.Inject
